Moved generate_basin_masks() to om4common
